refactor: move RadioConfig logic into separate ViewModel
