+++ app/build.gradle
@@ -162,7 +162,7 @@ dependencies {
 
     // Hilt
     implementation "com.google.dagger:hilt-android:$hilt_version"
-    // implementation "androidx.hilt:hilt-navigation-compose:1.0.0"
+    implementation "androidx.hilt:hilt-navigation-compose:1.0.0"
     ksp "com.google.dagger:hilt-compiler:$hilt_version"
 
     // Navigation
